Slowly move to a pure nixpkgs.
We still depend on <nixpkgs> but only exactly on _one place_. We can
soon swap it out with a pinned version of nixpkgs + an update script,
getting rid of all mutable state

Nixos configurations are just packages, see overlays/deployments.nix

deploy-remote.sh contains an example of how we can deploy to remote hosts
deploy-local.sh contains an example of how to deploy to a local host

We should probably make it act similar to nixos-rebuild, in that
we can just say  --targetHost=<blah> but that's something for next
time.

However, this shows that its totally feasible to have a totally pure
deployment model for both remote and local machine!