WIP: systemd-initrd
Start on an alternative initrd (Disables the NixOS initrd module)

The NixOS initrd is built up through a whole bunch of imperative
commands; and quite a bit of modules seem to depend on it. All those are
disabled for now (The ones I needed to get my own config going) and
would need either reimplemntation or changes to systemd declarative
primitives I think.

This commit can be tested with

nix run .#nixosConfigurations.t490s.config.system.build.runvm

It will run a QEMU VM with the systemd initrd. Nothing interesting
happens yet as bits and pieces are missing.

We'll work towards being able to boot my t490s (Which depends on LUKS)