t490s:  luks by uuid
So that we can use the options in systemd-cryptsetup-generator

Actually this device's disk adheres to  the format in gpt-auto-generator
so systemd should be able to figure this all out without _any_ settings.
but that's for another time. I want to keep my config compatible with
old and new world for now