WIP: systemd-initrd (#46)
* WIP: systemd-initrd

Start on an alternative initrd (Disables the NixOS initrd module)

The NixOS initrd is built up through a whole bunch of imperative
commands; and quite a bit of modules seem to depend on it. All those are
disabled for now (The ones I needed to get my own config going) and
would need either reimplemntation or changes to systemd declarative
primitives I think.